0. She locked away a secret, deep inside herself, something she once knew to be true... but chose to forget

+++ 0x08-recursion/0-puts_recursion.c
@@ -15,6 +15,6 @@ return;
 }
 _putchar(*s);
 _putchar('\n');
-_puts_recursion(*s ++);
+_puts_recursion((int) *s ++);
 
 }
